fix: keep indent by default to partial expression #482
