+++ src/template.rs
@@ -389,6 +389,7 @@ impl Template {
         template_stack: &mut VecDeque<Template>,
         source: &str,
         current_span: &Span<'_>,
+        prevent_indent: bool,
     ) -> bool {
         let with_trailing_newline = grammar::starts_with_empty_line(&source[current_span.end()..]);
 
@@ -396,7 +397,9 @@ impl Template {
             let with_leading_newline =
                 grammar::ends_with_empty_line(&source[..current_span.start()]);
 
-            if with_leading_newline {
+            // prevent_indent: a special toggle for partial expression
+            // (>) that leading whitespaces are kept, default to false
+            if prevent_indent && with_leading_newline {
                 let t = template_stack.front_mut().unwrap();
                 // check the last element before current
                 if let Some(RawString(ref mut text)) = t.elements.last_mut() {
@@ -602,6 +605,7 @@ impl Template {
                             &mut template_stack,
                             source,
                             &span,
+                            true,
                         );
 
                         let t = template_stack.front_mut().unwrap();
@@ -623,6 +627,7 @@ impl Template {
                             &mut template_stack,
                             source,
                             &span,
+                            true,
                         );
 
                         let t = template_stack.pop_front().unwrap();
@@ -678,12 +683,15 @@ impl Template {
                                 t.push_element(el, line_no, col_no);
                             }
                             Rule::decorator_expression | Rule::partial_expression => {
-                                // standalone statement check, it also removes leading whitespaces of
-                                // previous rawstring when standalone statement detected
+                                // do not auto trim ident spaces for
+                                // partial_expression(>)
                                 trim_line_required = Template::process_standalone_statement(
                                     &mut template_stack,
                                     source,
                                     &span,
+                                    // preventIndent set to false when
+                                    // rule is partial_expression
+                                    !(rule == Rule::partial_expression),
                                 );
 
                                 let decorator = DecoratorTemplate {
@@ -707,6 +715,7 @@ impl Template {
                                     &mut template_stack,
                                     source,
                                     &span,
+                                    true,
                                 );
 
                                 let mut h = helper_stack.pop_front().unwrap();
@@ -737,6 +746,7 @@ impl Template {
                                     &mut template_stack,
                                     source,
                                     &span,
+                                    true,
                                 );
 
                                 let mut d = decorator_stack.pop_front().unwrap();
@@ -768,6 +778,7 @@ impl Template {
                             &mut template_stack,
                             source,
                             &span,
+                            true,
                         );
 
                         let text = span
@@ -782,6 +793,7 @@ impl Template {
                             &mut template_stack,
                             source,
                             &span,
+                            true,
                         );
 
                         let text = span
